jgmenu-config.py: simplify 'amend'
  - Add jgmenurc default values as dict to avoid reading external file

  - Append missing key/value pairs instead of inserting in accordance
    with template. This significantly simplies and speeds up the code.

  - Add --dryrun option to just check for missing items, but without
    modifying jgmenurc

+++ noncore/config/jgmenu-config.py
@@ -8,24 +8,82 @@
 import argparse
 import sys
 import os
-import subprocess
-import re
 
 DEFAULT_CONFIG_FILE = "~/.config/jgmenu/jgmenurc"
+
+def jgmenurc():
+    """ return config file keys and values """
+    keys_and_values = {
+        "stay_alive": "1",
+        "hide_on_startup": "0",
+        "csv_cmd": "pmenu",
+        "tint2_look": "1",
+        "at_pointer": "0",
+        "terminal_exec": "x-terminal-emulator",
+        "terminal_args": "-e",
+        "monitor": "0",
+        "hover_delay": "100",
+        "hide_back_items": "1",
+        "columns": "1",
+        "tabs": "120",
+        "menu_margin_x": "0",
+        "menu_margin_y": "0",
+        "menu_width": "200",
+        "menu_height_min": "0",
+        "menu_height_max": "0",
+        "menu_height_mode": "static",
+        "menu_padding_top": "5",
+        "menu_padding_right": "5",
+        "menu_padding_bottom": "5",
+        "menu_padding_left": "5",
+        "menu_radius": "1",
+        "menu_border": "0",
+        "menu_halign": "left",
+        "menu_valign": "bottom",
+        "sub_spacing": "1",
+        "sub_padding_top": "auto",
+        "sub_padding_right": "auto",
+        "sub_padding_bottom": "auto",
+        "sub_padding_left": "auto",
+        "sub_hover_action": "1",
+        "item_margin_x": "3",
+        "item_margin_y": "3",
+        "item_height": "25",
+        "item_padding_x": "4",
+        "item_radius": "1",
+        "item_border": "0",
+        "item_halign": "left",
+        "sep_height": "5",
+        "sep_halign": "left",
+        "sep_markup": "",
+        "font": "",
+        "font_fallback": "xtg",
+        "icon_size": "22",
+        "icon_text_spacing": "10",
+        "icon_theme": "",
+        "icon_theme_fallback": "xtg",
+        "arrow_string": "▸",
+        "arrow_width": "15",
+        "color_menu_bg": "#000000 85",
+        "color_menu_border": "#eeeeee 8",
+        "color_norm_bg": "#000000 00",
+        "color_norm_fg": "#eeeeee 100",
+        "color_sel_bg": "#ffffff 20",
+        "color_sel_fg": "#eeeeee 100",
+        "color_sel_border": "#eeeeee 8",
+        "color_sep_fg": "#ffffff 20",
+        "color_scroll_ind": "#eeeeee 40",
+        "csv_name_format": "%n (%g)",
+        "csv_single_window": "0",
+        "csv_no_dirs": "0",
+        "csv_i18n": ""
+        }
+    return keys_and_values
 
 def die(msg):
     """ exit with message """
     sys.stderr.write("fatal: {}\n".format(msg))
     sys.exit(1)
-
-def read_pipe(cmd):
-    """ return stdout from subprocess """
-    proc = subprocess.Popen(cmd, stdout=subprocess.PIPE,
-                            stderr=subprocess.PIPE, shell=True)
-    (out, err) = proc.communicate()
-    if proc.returncode != 0:
-        die(err)
-    return out
 
 def get_key(line):
     """ return key from "key = value" string """
@@ -45,53 +103,10 @@ def get_keys(lines):
             keys.append(s)
     return keys
 
-def missing_keys(template_lines, conf_lines):
-    """ compare {config,template} files and return missing keys """
-    template_keys = get_keys(template_lines)
+def get_missing_keys(conf_lines):
+    """ compare {config,template} lists and return missing keys """
     conf_keys = get_keys(conf_lines)
-    return [item for item in template_keys if item not in set(conf_keys)]
-
-def nr_comment_lines_at_start(conf_lines):
-    """ calculate number of non-key/value lines at start of file """
-    i = 0
-    for line in conf_lines:
-        if "=" in line:
-            break
-        else:
-            i = i + 1
-    return i
-
-def key_in_line(key, line):
-    """ acurately determie if key exists in key/value line """
-    return bool(re.search(r'[#\s]*' + key + r'\s*=.*', line))
-
-def key_in_conf_lines(key, conf_lines):
-    """ check if keys exists in conf file """
-    for line in conf_lines:
-        if key_in_line(key, line):
-            return True
-    return False
-
-def insert_pos(template_lines, conf_lines, key):
-    """ find position to insert new line in accordance with template """
-    found_key = False
-    line_to_add = ''
-    for t_line in reversed(template_lines):
-        if "=" not in t_line:
-            continue
-        if found_key:
-            prev_key = get_key(t_line)
-            # Consider the case where several conf_file items are missing
-            if not key_in_conf_lines(prev_key, conf_lines):
-                continue
-            for c_index, c_line in enumerate(conf_lines):
-                if key_in_line(prev_key, c_line):
-                    return int(c_index) + 1, line_to_add
-        if get_key(t_line) == key:
-            line_to_add = t_line
-            found_key = True
-    # we'll onle get here if it's the first key/value line
-    return nr_comment_lines_at_start(conf_lines), line_to_add
+    return [item for item in jgmenurc() if item not in set(conf_keys)]
 
 def resolve(filename):
     """ expand ~ and $foo """
@@ -103,64 +118,24 @@ def resolve(filename):
         die("file '{}' does not exist".format(filename))
     return filename
 
-def get_template_filename():
-    """ get filename for template jgmenurc config file """
-    s = read_pipe("jgmenu_run --exec-path").strip().decode("utf-8")
-    s += "/jgmenurc"
-    if not os.path.exists(s):
-        die("$libexecdir/jgmenurc does not exist")
-    return s
-
-def adjust_line_spacing(conf_lines, template_lines):
-    """ adjust line spacing in accordance with template """
-    keys_with_space_after = []
-    space = False
-    for line in reversed(template_lines):
-        if line == '':
-            space = True
-        elif "=" in line and space:
-            keys_with_space_after.append(get_key(line))
-            space = False
-    new_list = []
-    first_key_line = True
-    for line in conf_lines:
-        if line == '':
-            continue
-        if '=' in line and first_key_line:
-            new_list.append('')
-            first_key_line = False
-        new_list.append(line)
-        if '=' in line and get_key(line) in keys_with_space_after:
-            new_list.append('')
-    return new_list
-
-def amend(conf_filename):
-    """ amend config-file based on template """
+def amend(conf_filename, isdryrun):
+    """ amend config-file with missing items """
     conf_filename = resolve(conf_filename)
-    template_filename = get_template_filename()
     with open(conf_filename, "r+") as f:
         conf_lines = f.read().split("\n")
-    with open(template_filename, "r") as f:
-        template_lines = f.read().split("\n")
 
-    keys = missing_keys(template_lines, conf_lines)
-    if keys:
-        print("info: add keys {}".format(keys))
-    else:
+    missing_keys = get_missing_keys(conf_lines)
+    if not missing_keys:
         return
 
-    # using insert_pos(), we may not get the correct line spacing
-    for key in keys:
-        i_pos, line = insert_pos(template_lines, conf_lines, key)
-        conf_lines.insert(i_pos, line)
+    if isdryrun:
+        print("info: missing keys {}".format(missing_keys))
+        return
 
-    conf_lines = adjust_line_spacing(conf_lines, template_lines)
+    print("info: add keys {}".format(missing_keys))
+    for key in missing_keys:
+        conf_lines.append("#{} = {}".format(key, jgmenurc()[key]))
 
-    # delete empty lines at end of file
-    for line in reversed(conf_lines):
-        if line == '':
-            conf_lines.pop()
-            break
     # write config file
     with open(conf_filename, "w") as f:
         for line in conf_lines:
@@ -175,9 +150,11 @@ def main():
                                          with missing items")
     amend_parser.add_argument("--file", metavar="FILE", help="specify config \
                               file", default=DEFAULT_CONFIG_FILE)
+    amend_parser.add_argument("--dryrun", action="store_true",
+                              help="list missing items, but do not amend file")
     args = parser.parse_args()
     if args.command == "amend":
-        amend(args.file)
+        amend(args.file, args.dryrun)
     else:
         parser.print_help(sys.stderr)
